Add compas View

+++ app/src/main/java/com/alastor/compassproject/CompassCallback.kt
@@ -2,5 +2,5 @@ package com.alastor.compassproject
 
 interface CompassCallback {
 
-    fun onSensorChanged(degree: Int, azimuth: Float)
+    fun onSensorChanged(degree: Float, azimuth: Float)
 }
